Add a class to the div and image and style it
Added the class, "home-recipe-container" to the div
element for jollofrice and add the class,
"home-image" to the jollofrice image in the home
page. Made the display of the div inline-block and
made the width and height 200px to fit the link
and image. Also, the inline-block display will
make the other links fit on the same line when
their containers and images are added.